abstract test case for use in testing acions plugins
